add a custom user model and manager
The custom user model and manager exist in authservice app. We need to
cusomtise how Users are validated, we have to extend the data on users,
we have to change how user passwords are created, and there will
probably be more. Having the custom user allows us to easily tick of all
requirements.